requirements.txt
fastapi==0.115.5
uvicorn==0.35.0,<1
python-dotenv==1.0.1
redis==5.0.8
redisvl==0.3.0
pydantic==2.11.3,<3
PyJWT==2.9.0
httpx==0.28.1,<1
sentence-transformers==3.2.1
gunicorn==22.0.0